requirements.txt
boto3==1.34.28 

pandas==2.1.4 
pyarrow==14.0.2 
pydantic==2.5.3 
python-dotenv==1.0.0 
soundfile==0.12.1
tqdm==4.66.1
apache-airflow
